+++ api.php
@@ -1,119 +1,108 @@
 <?php
-    $api = true;
-    require "scripts/pi-hole/php/password.php";
-    require "scripts/pi-hole/php/auth.php";
+	error_reporting(E_ALL);
+	require("scripts/pi-hole/php/FTL.php");
+	$socket = connectFTL("127.0.0.1", 4711);
+	header('Content-type: application/json');
 
-    check_cors();
+	$data = [];
 
-    include('scripts/pi-hole/php/data.php');
-    header('Content-type: application/json');
+	if (isset($_GET['summary']) || isset($_GET['summaryRaw']))
+	{
+		sendRequestFTL("stats");
+		$return = getResponseFTL();
 
-    $data = array();
+		$stats = [];
+		foreach($return as $line)
+		{
+			$tmp = explode(" ",$line);
 
-    // Non-Auth
+			if(isset($_GET['summary']))
+			{
+				if($tmp[0] !== "ads_percentage_today")
+				{
+					$stats[$tmp[0]] = number_format($tmp[1]);
+				}
+				else
+				{
+					$stats[$tmp[0]] = number_format($tmp[1], 1, '.', '');
+				}
+			}
+			else
+			{
+				$stats[$tmp[0]] = $tmp[1];
+			}
+		}
+		$data = array_merge($data,$stats);
+	}
 
-    if (isset($_GET['summaryRaw'])) {
-        $data = array_merge($data,  getSummaryData());
-    }
+	if (isset($_GET['overTimeData10mins']))
+	{
+		sendRequestFTL("overTime");
+		$return = getResponseFTL();
 
-    if (isset($_GET['summary']) || !count($_GET)) {
-        $sum = getSummaryData();
-        $sum['ads_blocked_today'] = number_format( $sum['ads_blocked_today']);
-        $sum['dns_queries_today'] = number_format( $sum['dns_queries_today']);
-        $sum['ads_percentage_today'] = number_format( $sum['ads_percentage_today'], 1, '.', '');
-        $sum['domains_being_blocked'] = number_format( $sum['domains_being_blocked']);
-        $data = array_merge($data,  $sum);
-    }
+		$domains_over_time = [];
+		$ads_over_time = [];
+		foreach($return as $line)
+		{
+			$tmp = explode(" ",$line);
+			$domains_over_time[] = $tmp[1];
+			$ads_over_time[] = $tmp[2];
+		}
+		$result = ['domains_over_time' => $domains_over_time,
+		           'ads_over_time' => $ads_over_time];
+		$data = array_merge($data, $result);
+	}
 
-    if (isset($_GET['overTimeData'])) {
-        $data = array_merge($data,  getOverTimeData());
-    }
+	if (isset($_GET['topItems']))
+	{
+		sendRequestFTL("top-domains");
+		$return = getResponseFTL();
+		$top_queries = [];
+		foreach($return as $line)
+		{
+			$tmp = explode(" ",$line);
+			$top_queries[$tmp[2]] = $tmp[1];
+		}
 
-    if (isset($_GET['overTimeData10mins'])) {
-        $data = array_merge($data,  getOverTimeData10mins());
-    }
+		sendRequestFTL("top-ads");
+		$return = getResponseFTL();
+		$top_ads = [];
+		foreach($return as $line)
+		{
+			$tmp = explode(" ",$line);
+			$top_ads[$tmp[2]] = $tmp[1];
+		}
 
-    // Auth Required
-    $auth = true;
+		$result = ['top_queries' => $top_queries,
+		           'top_ads' => $top_ads];
 
-    if (isset($_GET['topItems'])) {
-        $data = array_merge($data,  getTopItems($_GET['topItems']));
-    }
+		$data = array_merge($data, $result);
+	}
 
-    if (isset($_GET['recentItems']) && $auth) {
-        if (is_numeric($_GET['recentItems'])) {
-            $data = array_merge($data,  getRecentItems($_GET['recentItems']));
-        }
-    }
+	if (isset($_GET['topClients']) || isset($_GET['getQuerySources']))
+	{
+		sendRequestFTL("top-clients");
+		$return = getResponseFTL();
+		$top_clients = [];
+		foreach($return as $line)
+		{
+			$tmp = explode(" ",$line);
+			if(count($tmp) == 4)
+			{
+				$top_clients[$tmp[3]."|".$tmp[2]] = $tmp[1];
+			}
+			else
+			{
+				$top_clients[$tmp[2]] = $tmp[1];
+			}
+		}
 
-    if (isset($_GET['getQueryTypes']) && $auth) {
-        $data = array_merge($data, getIpvType());
-    }
+		$result = ['top_sources' => $top_clients];
+		$data = array_merge($data, $result);
+	}
 
-    if (isset($_GET['getForwardDestinations']) && $auth) {
-        $data = array_merge($data, getForwardDestinations());
-    }
+	echo json_encode($data);
 
-    if (isset($_GET['getQuerySources']) && $auth) {
-        $data = array_merge($data, getQuerySources());
-    }
+	disconnectFTL();
 
-    if (isset($_GET['getAllQueries']) && $auth) {
-        $data = array_merge($data, getAllQueries($_GET['getAllQueries']));
-    }
-
-    if (isset($_GET['enable'], $_GET['token']) && $auth) {
-        check_csrf($_GET['token']);
-        exec('sudo pihole enable');
-        $data = array_merge($data, Array(
-            "status" => "enabled"
-        ));
-    }
-    elseif (isset($_GET['disable'], $_GET['token']) && $auth) {
-        check_csrf($_GET['token']);
-        $disable = intval($_GET['disable']);
-        // intval returns the integer value on success, or 0 on failure
-        if($disable > 0)
-        {
-            exec("sudo pihole disable ".$disable."s");
-        }
-        else
-        {
-            exec('sudo pihole disable');
-        }
-        $data = array_merge($data, Array(
-            "status" => "disabled"
-        ));
-    }
-
-    if (isset($_GET['getGravityDomains'])) {
-        $data = array_merge($data, getGravity());
-    }
-
-    if (isset($_GET['tailLog'])) {
-        $data = array_merge($data, tailPiholeLog($_GET['tailLog']));
-    }
-
-    function filterArray(&$inArray) {
-	    $outArray = array();
-	    foreach ($inArray as $key=>$value) {
-	        if (is_array($value)) {
-	            $outArray[htmlspecialchars($key)] = filterArray($value);
-            } else {
-	            $outArray[htmlspecialchars($key)] = htmlspecialchars($value);
-            }
-	    }
-	    return $outArray;
-    }
-
-    $data = filterArray($data);
-
-    if(isset($_GET["jsonForceObject"]))
-    {
-        echo json_encode($data, JSON_FORCE_OBJECT);
-    }
-    else
-    {
-        echo json_encode($data);
-    }
 ?>
